Adapt unit tests to refactors made
